Initial commit.
Adding new file. This script can be used to assist in a speedier backup process of multiple applications to a USB drive. It expects a USB drive - if no USB drive is plugged into the computer, the drive option simply becomes a standard text input box; otherwise it's a select box. If the (compiled) script is run from a USB drive, that drive should default to the selected option.